extract flask filters and context_processor into separated files
